0326 commit

+++ classes/product.php
@@ -4,9 +4,9 @@ require_once "database.php";
 
 class Product extends Database{
 
-    //Get All Products
+    //Get All Products (index.php)
     public function getProducts(){
-        $sql = "SELECT * FROM products ORDER BY products.product_name ASC";
+        $sql = "SELECT * FROM products WHERE product_stock != 0 ORDER BY products.product_name ASC";
         //$result = $this->conn->query($sql);
         //print_r($result);
 
@@ -19,22 +19,22 @@ class Product extends Database{
         }
     }
 
-    //Admin Page editProduct Page
-    public function getProduct($product_id){
-        $sql = "SELECT * FROM products WHERE product_id = '$product_id'";
+    public function getProductsAdmin(){
+        $sql = "SELECT * FROM products ORDER BY products.product_name ASC";
 
         if($result = $this->conn->query($sql)){
-            //print_r($result);
-            return $result->fetch_assoc();
+            return $result;
+            //return $result->fetch_assoc();
             exit;
         }else{
-            die("Error retrieving product: " . $this->conn->error);
+            die("Error retrieving products: " . $this->conn->error);
         }
+
     }
 
     public function searchProduct($keyword){
-        $sql = "SELECT * FROM products WHERE product_name like '%$keyword%'";
-        //print_r($sql);
+        
+        $sql = "SELECT * FROM products WHERE product_name LIKE '%$keyword%' AND product_stock != 0";
 
         if($result = $this->conn->query($sql)){
             //print_r($result);
@@ -46,9 +46,35 @@ class Product extends Database{
         }
     }
 
-    public function addProduct($img_url, $product_name, $unit_price, $product_stock,$tmp_name){
-        $sql = "INSERT INTO products(product_name, unit_price, product_stock, img_url) 
-        VALUES('$product_name','$unit_price','$product_stock','$img_url')";
+    public function calculateProductStock($customer_id){
+        $sql1 = "SELECT carts.product_id, carts.quantity, products.product_stock FROM carts 
+        INNER JOIN products ON carts.product_id = products.product_id
+        WHERE customer_id = '$customer_id'";
+        
+        if($result = $this->conn->query($sql1)){
+
+            while ($row = $result->fetch_assoc()){
+                print_r($row);
+
+                $product_id = $row['product_id'];
+
+                $new_product_stock = $row['product_stock'] - $row['quantity'];
+                //print_r($new_product_stock);
+
+                $sql2 = "UPDATE products SET product_stock = '$new_product_stock' WHERE product_id = '$product_id'";
+                print_r($sql2);
+                //UPDATE products SET product_stock = 80 WHERE product_id = 1
+
+                $this->conn->query($sql2);
+
+            }
+        }
+    }
+
+    //Admin Page
+    public function addProduct($img_url, $product_name, $unit_price, $product_stock,$tmp_name, $description){
+        $sql = "INSERT INTO products(product_name, unit_price, product_stock, img_url, `description`) 
+        VALUES('$product_name','$unit_price','$product_stock','$img_url', '$description')";
 
         if ($this->conn->query($sql)) {
             $destination = "../img/" . basename($img_url);
@@ -65,21 +91,40 @@ class Product extends Database{
          }
     }
 
-    public function updateProduct($product_id, $product_name, $unit_price, $product_stock, $img_url, $tmp_name){
+    //Admin Page editProduct Page
+    public function getProduct($product_id){
+        $sql = "SELECT * FROM products WHERE product_id = '$product_id'";
+
+        if($result = $this->conn->query($sql)){
+            //print_r($result);
+            return $result->fetch_assoc();
+            exit;
+        }else{
+            die("Error retrieving product: " . $this->conn->error);
+        }
+    }
+
+    public function updateProduct($product_id, $product_name, $unit_price, $product_stock, $img_url, $tmp_name, $description){
         $sql = "UPDATE products 
         SET product_name = '$product_name', unit_price = '$unit_price',
-        product_stock = '$product_stock', img_url = '$img_url'
+        product_stock = '$product_stock', img_url = '$img_url', `description` = '$description'
         WHERE product_id = '$product_id'";
 
         if ($this->conn->query($sql)) {
             $destination = "../img/" . basename($img_url);
-            if (move_uploaded_file($tmp_name, $destination)) {
-                //move_uploaded_file(from, to)
 
-            header("location: ../views/admin.php");
-            exit;
-            } else {
-            die("Error moving the photo.");
+            if((file_exists($destination))){
+                header("location: ../views/admin.php?page=1");
+                exit;
+            }else{
+                
+                if (move_uploaded_file($tmp_name, $destination)) {
+                    //move_uploaded_file(from, to)
+                header("location: ../views/admin.php?page=1");
+                exit;
+                } else {
+                    die("Error moving the photo.". $this->conn->error);
+                }
             }
         } else {
             die("Error uploading photo: " . $this->conn->error);
@@ -96,5 +141,5 @@ class Product extends Database{
             die("Error deleteing Product " . $this->conn->error);
         }
     }
-    
+
 }
